+++ Scripts/Components/Player/DSC_ADD_Player.cs
@@ -1,0 +1,9 @@
+﻿using Unity.Entities;
+
+namespace DSC.Actor.DOTS
+{
+    public struct DSC_ADD_Player : IComponentData
+    {
+        public int m_nID;
+    }
+}